backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose
email-validator
bcrypt==4.0.1
